requirements.txt
requests==2.31.0
loguru==0.7.2
apscheduler==3.10.4
python-dotenv==1.0.1
kafka-python
fastapi
uvicorn
